# schemas/api/openapi.yaml
openapi: 3.1.0
info:
  title: ShakerProxy Control API
  version: 0.1.0-dev
  description: |
    Setup-safe API. Only operations implemented by this build are listed.

    Discover resources with GET /api/v1 and fetch this document from GET /api/v1/openapi.yaml.
    Every error, including unknown routes (404 not_found) and unsupported methods
    (405 method_not_allowed with an Allow header), uses the JSON shape
    {"error":{"code":"snake_case","message":"What went wrong and what to do next."}}.
    Request bodies are JSON (Content-Type application/json, optional charset=utf-8);
    decoding errors name the offending field.

    Authentication: sign in with POST /auth/login and send Authorization: Bearer <session_token>.
    Sessions slide (1 hour idle, 12 hours maximum); authenticated responses carry
    X-ShakerProxy-Session-Expires-At. Scoped API tokens (POST /auth/tokens) use the same header.

    Password confirmation: fields named "password" confirm the administrator. After any
    successful password check (including sign-in) the same session may omit "password" for
    10 minutes on routes marked "recent confirmation"; they answer 401 reauthentication_required
    when a password is needed. Routes marked "password every time" (network commit/confirm,
    deletion execution and retention, API token creation, enabling HTTPS decryption for every
    client, enabling decrypted HTTP body retention) answer 401 password_required without one.
servers:
  - url: /api/v1
paths:
...
  /traffic-policy:
    get:
      operationId: getTrafficPolicy
      description: Session-only current DNS and TLS traffic-policy document from the privileged host. The installed production gateway profile is required.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Current revisioned traffic policy, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyDocument"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    put:
      operationId: applyTrafficPolicy
      description: Freshly reauthenticated optimistic apply through the privileged gateway. Active controls require a confirmed IPv4 routed or single-arm plan, and listener failure leaves the packet path fail-open.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ApplyTrafficPolicyRequest"}}}}
      responses:
        "200": {description: Applied revisioned traffic policy, content: {application/json: {schema: {$ref: "#/components/schemas/TrafficPolicyDocument"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /traffic-policy/bypass:
    post:
      operationId: addTLSBypass
      description: One-click "bypass this host" for a TLS failure. With device (ID, MAC, IP or name) the host passes through undecrypted for that device only; without it, for every device. Session or lab:write token; no password. Applied through the revisioned, audited traffic-policy path.
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/TLSBypassRequest"}}}}
      responses:
        "200": {description: Bypass in effect from the next connection, content: {application/json: {schema: {$ref: "#/components/schemas/TLSBypassResult"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    delete:
      operationId: removeTLSBypass
      description: Removes a host bypass added with POST, for one device or for every device. Session or lab:write token.
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: host, required: true, schema: {type: string, maxLength: 253}}
        - {in: query, name: device, schema: {type: string, maxLength: 256}, description: Device ID, MAC, IP or friendly name; omit for the every-device bypass.}
      responses:
        "200": {description: Bypass removed, content: {application/json: {schema: {$ref: "#/components/schemas/TLSBypassResult"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
  /interception-ca/onboarding:
    get:
      operationId: getInterceptionCAOnboarding
      description: Where lab devices get the interception CA (http://<lab gateway>/, served only while HTTPS decryption is configured) and per-platform install steps and limitations. When the page is not reachable, available is false and reason says what to do next. Session or system:read token.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Onboarding status, content: {application/json: {schema: {$ref: "#/components/schemas/InterceptionCAOnboarding"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
...
  /devices/{deviceID}/controls:
    parameters:
      - {in: path, name: deviceID, required: true, schema: {type: string, minLength: 1, maxLength: 128}, description: "Device ID, MAC (any separator), current IP or friendly name. Ambiguous references return 409 device_ambiguous with candidates; unknown ones 404 device_not_found."}
    get:
      operationId: getDeviceControls
      description: The device's lab controls (Decrypt HTTPS, internet, blocked domains) and whether they are enforced right now, with plain-language notes. Session or devices:read token.
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Current device controls, content: {application/json: {schema: {$ref: "#/components/schemas/DeviceControls"}}}}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
    put:
      operationId: putDeviceControls
      description: "Changes any of the given controls; omitted fields keep their value. decrypt_https adds the device to HTTPS decryption for selected devices (turning it on when the CA exists) and blocks QUIC for it; internet BLOCK drops the device's traffic leaving the lab (matched by MAC, falling back to current IPs); blocked_domains answer NXDOMAIN, subdomains included, through ShakerProxy DNS. Revisioned and audited through the traffic-policy apply path. Session or lab:write token; no password."
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/DeviceControlsUpdate"}}}}
      responses:
        "200": {description: Controls after the change, content: {application/json: {schema: {$ref: "#/components/schemas/DeviceControls"}}}}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, description: Browser sessions authenticate every route. Expiring scoped API tokens authenticate only explicitly scoped read/automation routes and cannot manage credentials, forwarders, networking, deletion, exports, rules, or trust material.}
  responses:
    Error:
      description: Structured error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string, pattern: "^[a-z0-9_]+$"}
                  message: {type: string, description: What went wrong and what to do next.}
...
  schemas:
...
    InterceptionCAOnboarding:
      type: object
      additionalProperties: false
      required: [schema, available, reason, sha256_fingerprint, common_name, not_after, urls, instructions]
      properties:
        schema: {const: 1}
        available: {type: boolean, description: True when lab devices can open the onboarding page now.}
        reason: {type: string, description: Empty when available; otherwise what to do next.}
        sha256_fingerprint: {type: string, description: "Upper-case colon-separated SHA-256, empty when no CA exists."}
        common_name: {type: string}
        not_after: {type: string, description: RFC 3339 expiry, empty when no CA exists.}
        urls:
          type: array
          maxItems: 2
          items:
            type: object
            additionalProperties: false
            required: [label, url]
            properties:
              label: {type: string}
              url: {type: string}
        instructions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [platform, title, steps, limitations]
            properties:
              platform: {type: string, enum: [ios, android, android-tv, macos, windows, linux, smart-tv, other]}
              title: {type: string}
              steps: {type: array, items: {type: string}}
              limitations: {type: array, items: {type: string}}
...
    TLSBypassRequest:
      type: object
      additionalProperties: false
      required: [host]
      properties:
        host: {type: string, maxLength: 253}
        device: {type: string, maxLength: 256, description: Device ID, MAC, IP or friendly name; omit to bypass for every device.}
    TLSBypassResult:
      type: object
      additionalProperties: false
      required: [schema, host, device_id, scope, revision, notes]
      properties:
        schema: {const: 1}
        host: {type: string}
        device_id: {type: [string, "null"]}
        scope: {type: string, enum: [device, all_devices]}
        removed: {type: boolean}
        revision: {type: integer, minimum: 1}
        notes: {type: array, items: {type: string}}
...
    TrafficPolicyDocument:
      type: object
      additionalProperties: false
      required: [schema, policy, digest, applied_at]
      properties:
        schema: {const: 1}
        policy: {$ref: "#/components/schemas/TrafficPolicy"}
        digest: {type: string, pattern: "^[a-f0-9]{64}$"}
        applied_at: {type: string, format: date-time}
        previous: {$ref: "#/components/schemas/TrafficPolicy"}
...
    ApplyTrafficPolicyRequest:
      type: object
      additionalProperties: false
      description: password is needed every time the policy turns on HTTPS decryption for every client (tls_interception.enabled with no selected_device_ids) while it was not already on; otherwise a recent confirmation is enough.
      required: [expected_revision, policy]
      properties:
        expected_revision: {type: integer, minimum: 1}
        policy: {$ref: "#/components/schemas/TrafficPolicy"}
        password: {type: string, minLength: 1, maxLength: 1024}
...